Use pullSecret in builds that access registry.redhat.io
Currently our SSO build needs to access registry.redhat.io, this
requires authentication in order to complete the pull. To do this
we need to reference a pullSecret in the BuildConfig. As pullSecret
is a LocalObjectReference we need to copy the container registry
credentials secret into the NAPS namespace.

Verification:
- Run the installer
- Ensure the SSO build completes successfully